refactor : Lotto, LottoNumber  네이밍 변경
