--- src/main/kotlin/lotto/domain/LottoNumber.kt
@@ -1,3 +1,0 @@
-package lotto.domain
-
-class LottoNumber(val value: List<Int>)
