+++ src/main/kotlin/lotto/domain/strategy/LottoFactory.kt
@@ -1,10 +1,10 @@
 package lotto.domain.strategy
 
-import lotto.domain.Lotto
+import lotto.domain.Lottos
 
 class LottoFactory {
 
-    fun generate(lottoCount: Int, lottoGeneratorStrategy: LottoGeneratorStrategy): Lotto {
+    fun generate(lottoCount: Int, lottoGeneratorStrategy: LottoGeneratorStrategy): Lottos {
         return lottoGeneratorStrategy.generate(lottoCount)
     }
 }
